requirements210.txt
absl-py==1.4.0
astunparse==1.6.3
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.0.7
cycler==0.11.0
flatbuffers==2.0
fonttools==4.38.0
gast
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.7.0
idna==3.4
keras==2.10.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
numpy==1.23.5
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pillow==9.4.0
protobuf==3.19.6
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
setuptools==65.5.0
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorflow-cpu==2.10.1
termcolor==2.2.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wheel==0.38.4
wrapt==1.14.1
pyinstaller
